AppDetails file renames
